lee999 修改爱慕 地图定位的功能

+++ MySuperApp/Class/UI/mainpage/map/StoreDetail20ViewController.h
@@ -31,7 +31,7 @@
 
 @property (nonatomic, retain)StoresStores *store;
 
-@property (nonatomic, assign)int index;
+@property (nonatomic, assign)NSInteger index;
